Allow Windows nodejs to search <drive>:\node_modules
Currently, there is a bug where the nodeModulesPaths function will not properly return the node_modules directory in the root of the disk.  This is due to the following statement being true:

    path.join.apply(path, ["c:", "dev", "workspace"].slice(0, 1)) === "c:."

This resulted in the returned path being "c:node_modules", which is invalid.

This code change fixes the bug by doing the path.join in one statement, including "node_modules" as part of the arguments.