fix: allow overriding entry point in browser field
Some modules are [published](https://unpkg.com/cborg@1.1.2/package.json) with no `main`
field in `package.json` and no `/index.js`, instead using `exports` to define the entry
point for the module.

They sometimes specify an entry point for the browser using a `"."` or `"./"` key in
the `browser` field, but currently these are broken with browserify as it tries to
resolve `/index.js` when no `main` field is present.

The change here is to allow define the entry point for the module using the `browser`
field:

```json
{
  "browser": {
    ".": "path/to/index.js"
  }
}
```